Initial commit to use util.py module.
