+++ hop-deploy.py
@@ -14,6 +14,12 @@ import urllib.error
 import ssl
 import threading
 
+from util import (
+    print_colored,
+    periodic_status_update,
+    run_command
+)
+
 HOPSWORKS_LOGO = """
 ██╗  ██╗    ██████╗    ██████╗    ███████╗   ██╗    ██╗    ██████╗    ██████╗    ██╗  ██╗   ███████╗
 ██║  ██║   ██╔═══██╗   ██╔══██╗   ██╔════╝   ██║    ██║   ██╔═══██╗   ██╔══██╗   ██║ ██╔╝   ██╔════╝
@@ -310,28 +316,6 @@ class HopsworksInstaller:
             "blue"
         )
 
-def print_colored(message, color, **kwargs):
-    colors = {
-        "red": "\033[91m", "green": "\033[92m", "yellow": "\033[93m",
-        "blue": "\033[94m", "magenta": "\033[95m", "cyan": "\033[96m",
-        "white": "\033[97m", "reset": "\033[0m"
-    }
-    print(f"{colors.get(color, '')}{message}{colors['reset']}", **kwargs)
-
-def run_command(command, verbose=False):
-    if verbose:
-        print_colored(f"Running: {command}", "cyan")
-    try:
-        result = subprocess.run(
-            command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True
-        )
-        if verbose and result.stdout:
-            print(result.stdout)
-        if result.stderr:
-            print_colored(result.stderr, "yellow")
-        return result.returncode == 0, result.stdout, result.stderr
-    except Exception as e:
-        return False, "", str(e)
 
 def get_user_input(prompt, options=None):
     while True:
@@ -350,15 +334,6 @@ def get_deployment_environment():
         [str(i) for i in range(1, len(environments) + 1)]
     )
     return environments[int(choice) - 1]
-
-def periodic_status_update(stop_event, namespace):
-    while not stop_event.is_set():
-        cmd = f"kubectl get pods -n {namespace} --no-headers"
-        success, output, _ = run_command(cmd, verbose=False)
-        if success:
-            pod_count = len(output.strip().split('\n'))
-            print_colored(f"\rCurrent status: {pod_count} pods created", "cyan", end='')
-        time.sleep(10)  # Update every 10 seconds
 
 def get_license_agreement():
     print_colored("\nChoose a license agreement:", "blue")
